+++ test/Services/ChessEngineService.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using System.Text.RegularExpressions;
 
 namespace ChessDroid.Services
 {